[IMP] account_intrastat: Add Grouped lines for intrastat report
In the intrastat report, each account move is now grouped by country
(using norm ISO 3166), currency (norm ISO 4217) and commodity code. Each
line can be unfolded to show each individual account move with its
corresponding values. When only Arrival or Dispatch is selected, the
total value for each group is shown as well as a total line for the
total value of Arrivals or Dispatches.

This change was done because usually people and
organisation to which they must report this information aren't
interested in each individual move but rather to have an overall view of
how much of a certain commodity was ship to a certain country.
The previous report was too verbose. People still have the options to
unfold all the lines if they need it.

task-3605696

closes odoo/enterprise#56353

X-original-commit: bb67df2f37a65acd9f62a28d8a52fbd335ef903e
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>