[IMP] social_linkedin: allow to add / view images in comment
Purpose
=======
Allow inserting image in comment and viewing the attached image
in existing comments.

We needed to increment the API version in order to use the new image
endpoint, instead of the old asset endpoint. Since this change,
sometimes the API does not return author information, so we also need
to handle that case now.

Task-3095529

Part-of: odoo/enterprise#35170