[IMP] social_linkedin: allow to edit comments
Purpose
=======
Allow to edit LinkedIn comments, like we do for other social medias.

Technical
=========
The like summary does not work on the edit endpoint (it always returns
zero), so we send only the diff to the web client (so the message
updates, and we preserve other keys).

Task-3095529

Part-of: odoo/enterprise#35170