[IMP] account_reports: Test that the Balance Sheet is balanced.
This commit introduces a l10n test over the Balance Sheet reports.

For each CoA, a company is set-up, and all the available Balance Sheet
reports for the company are tested.
The test generates a journal entry with a line in every account that is
not `off_balance` and not in the NON_TESTED_ACCOUNTS global at the head
of the file (useful for accounts that are forced to be generated, but
should never be used, like 999999 in Belgium).
It then checks whether the Balance Sheet is balanced.
The test does this both for debits and for credits.

To work, the test needs to know, for each Balance Sheet report:
- the XMLID of the Total Assets line
- the XMLID of the Total Liabilities line
- the XMLID of the Total Equity line (if separate from Liabilities)
- the column name containing the balance (if different from 'balance')

These are specified in the REPORT_CONFIG global at the head of the file.

The test will hard-fail if an available Balance Sheet report is not
configured in the REPORT_CONFIG global. This makes it mandatory for new
Balance Sheets to add their config there.

Setting the global IDENTIFY_INCORRECT_ACCOUNTS to True switches the
behaviour of the test: it identifies the accounts that cause the
imbalance using a binary search. This is off by default for runbot.

closes odoo/enterprise#65863

Task: 3060790
X-original-commit: 1d51326abc8dd7a2b882d932d885f5a776a7efe1
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>