[IMP] planning: add colors to the grouped gantt view
In this commit, we add different colorings to the group
rows (in the grouped gantt view) to indicate whether
the planned hours are more (red), less (orange), or equal
(green)to the resource's working hours for that day.

We take into account flexible hours, time off and non-human
resources.

task-3071699

closes odoo/enterprise#38343

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>