[IMP] helpdesk: remove helpdesk ticket type
purpose:
Currently, we have 4 ways of categorizing tickets: helpdesk teams, ticket types,
tags, and property fields.
It seems overkill and adds clutter to the views of tickets needlessly.
Among all of these fields, ticket types seem to be the least useful/versatile as
it is a simple selection field that could easily be added with property fields.
We are thus removing ticket types.

task-3242226

Part-of: odoo/enterprise#38943