[IMP] timesheet_grid: improve ux
The main goal of this task is to give users visual clues so
that they can see at glance there is an abnormal situation to address.

Before this commit:
1) grid view: It makes no sense to encode more than 24h in one day.
We can thus safely assume that the user mistyped, but the user is not
alerted about it.

2) if an employee has timesheeted less than they should have,
the company is potentially losing money. It must then be rectified asap.
Same when doing overtime, it can mean that the company will have to
compensate the employee accordingly. It is also a clue for the manager
that an employee may be overloaded.

3) using _work_intervals_batch to get resources work interval doesn't take
into consideration their calendar.

In this commit:
1) Cells that are > 24h are highlighted in red when grouping by an employee
or when viewing the 'my timesheets' menu

2) We colored the cell of the total per day of each employee in the section.
- red when there are missing timesheets
- orange if there is overtime.

3)using _get_valid_work_intervals fixed the issue.

Related PRs:
Community: https://github.com/odoo/odoo/pull/124512

task-3054342

closes odoo/enterprise#33549

Related: odoo/odoo#124512
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>