[IMP] sale_subscription{,_stock}: support postpaid contracts
Before this commit, the sale_subscription module could only work with
products being invoiced at the beginning of the period. It works well for
service but stockable products which are delivered don't fit well in
this paradigm.

For these kinds of products, we want to invoice at the end of the month
and the invoiced quantity should be deducted from de delivered amount.

Taskid: 3071118
Part-of: odoo/enterprise#41796
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>