[IMP] planning: improve mail templates, add QoL features
This commit will change mail templates in planning to make emails
more understandable, specific and breathable. Since these emails are
not worth taking so much space in the database, they will now be
automatically deleted.

When sending a shift to employees, a wizard pops up to let the user add
a work email to employees that don't have one so they can receive their
shift. However, there was no option to do that when creating a slot for
a single employee ; this has now been changed. A form view now pops up
whenever creating a slot allowing you to set a work email for it,
if the selected employee has no work email.

task-3251693

closes odoo/enterprise#42215

Related: odoo/odoo#124415
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>