[FIX] knowledge: copy kanban stages when copying an article
When the user copies an article, the stages set on the article items of
the article will not be transposed to the new article. As a result, the
embedded view will not have any stage set. To fix that issue, we will
properly copy the stage of the article to the new article.

Step to reproduce the issue:
1. Create an article
2. Insert a kanban view with /kanban
3. Create a few stages
4. Click on the "Create a Copy" button of the dropdown item
=> The embedded view will not have any stage set

TO BE: The embedded view should have the same stages as the original article.

task-3349324

closes odoo/enterprise#54267

X-original-commit: acace5a5cd816f4fc1a896e62959690298f1d547
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>