[IMP] project_enterprise{_hr_skills},industry_fsm: search tasks according to assignees' skills
Before this commit, when the user would like to assign a user according to
a certain skill, he cannot, he has to know which user has that skill to assign
it.

This commit adds a new module to make the bridge between project_enterprise
and hr_skills to be able to search the users to display when we group the tasks
by assignees to have the users with a certain skill searched in the search view.

task-3349411

closes odoo/enterprise#42175

Related: odoo/odoo#130980
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>