[IMP] downloads_account: Add selection of journal in Document Action 'import bank statement'
For now we didn't chose the bank journal because we assume that only
coda file where push like this, and in the coda file we can find in
which journal to send it. But Coda is limited to BE and lot of user
import CSV file for bank.

Add the selection of the journal (not mandatory) for Document action,
and inject CSV in the journal selected. If no journal selected inject
the CSV in the first one of the list. If it's a coda file, keep as now
and don't take care of the journal selected.

task-3288472

closes odoo/enterprise#46482

Signed-off-by: William André (wan) <wan@odoo.com>