[IMP] sale_subscription: Improve upselling alert for subscription
As the SO for subscription are updated every recurrence, it's hard
to keep track of possible upselling opportunities.

This change update the sale upselling opportunity for subscription
to save more information regarding the opportunity and don't erase
old activities to keep a better record of what has been deliverd/invoiced.

task-id : 3323129

closes odoo/enterprise#69155

X-original-commit: 74e9e77b2769ae033e4b82f4174de53ba8df957c
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>