[ADD] delivery_ups_rest: New UPS API Integration
UPS rolled out their REST API with 0Auth credentials and as of June 2023
have stopped issuing new API access keys used by the older SOAP Integration.
and as of June 2024, all requests with the old auth will be rejected.

The new REST API Integration is added in the new module `delivery_ups_rest`
to allow users a period to transition to the new integration.

The old module `delivery_ups` is now deprecated and will be deleted in
2024 when it becomes obsolete.

taskId: 3349269
Github Issue: odoo/odoo#123446

closes odoo/enterprise#51593

X-original-commit: afe5fbb3122b28b0d4db000119629ec5711a690f
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Ahmed Khalaf (ahkh) <ahkh@odoo.com>