[IMP] planning : undo auto-plan action
Before this commit, when the user was triggering the auto-plan feature,
there was no way for the user to come back to the previous schedule.

After this commit, the success toast notification is modified in order
to include un "undo" button allowing the user to come back to the
previous configuration.

closes odoo/enterprise#42992

Taskid: 3186645
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>