[IMP] l10n_be_(hr_contract_salary)_(payroll) : Simplify wishlist car flow
The objective of this PR is to simplify the way employees and applicants pick their cars in the salary configurator :
1. We now make a clear distinction between existing cars as fleet.vehicle records and fleet.vehicle.models, the selection in the company car will now only contain existing cars.
2. Wishlist was renamed to "To Order"
3. The user can now do a secondary simulation of his salary with his wishlist car
4. The payroll officer has now the possibility to display or not display the "To Order" field in the salary configurator.

Added fields :
ordered_car_id : Once an offer has been fully signed by both the employee and the hr officer, if the employee selected a car model to order, that model will be created as a fleet.vehicle and be assigned to this field. When the new car is available, a new offer will be generated with this car as the default one.

Modified behavior:
new_car and new_car_model_id : This field is now a purely technical field, it will serve as information during the signing process to show the hr officer, which car model has been chosen to be ordered. Once an offer is fully signed, it will be emptied again and transfered to ordered_car_id.

task-3356387

closes odoo/enterprise#43666

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>