[IMP] planning,**: improve gantt progress bar display format
**=mrp_workorder, project_enterprise, web_gantt

In this commit, we enhanced the Gantt view progress bar by modifying
the display format. Previously, when the hours were not rounded,
it would show as '4h' instead of '4h30'. We have now updated it to
accurately represent the exact duration. Additionally, we changed
the format from '4 h / 38 h' to '4h / 38h' for improved clarity.

task-3178655

closes odoo/enterprise#43261

Signed-off-by: Bruno Boi (boi) <boi@odoo.com>