[FIX] project_enterprise: fix group_expand method for tasks gantt view
This PR will fix three different issues with the tasks gantt view's
group expand method. The first issue is that it did not include closed
task for the past period, so no empty line was displayed for the user.
The second issue is that it did not include closed task for the current
period, so no empty line was displayed for the user. The third issue is
that it included private tasks, which should not be the case.

To fix this, the _group_expand_user_ids method was changed to include
closed tasks in the domain calculation for the previous/current period,
and to exclude private tasks. Tests were also added.

task-3389422

closes odoo/enterprise#50550

X-original-commit: e7f1efa3ba84bb65c49522dc7e23b874794eece7
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>