[FIX] sale_planning: ensure shift unplanned has allocated_hours>0
Before this commit, when the user copies the shifts in the previous
week and then auto plan, an error occurs because a record is used
but it does no longer exist in the database. The reason is the shift
unplanned will be used in the auto plan feature to plan it to an
available resource but the problem is that shift has allocated_hours
equals to 0 and so it is removed in the process since it means
we have no longer something to plan for a specific SOL.

This commit recomputes the `planning_hours_planned` to be sure the
post_process, called in the create method, remove the shifts unplanned
with allocated hours equals to 0.

closes odoo/enterprise#50335

X-original-commit: da69234fd73744fb576153b8f466a57e28261b1d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>