[ADD] l10n_lt_intrastat: Add Lithuanian Intrastat report
This commit add the Lithuanian intrastat report. This comes with the possibility
to export that report into an XML file and also adds the country region code.
The XML export can than be submitted to the officiql intrastat reporting
of Lithuania: https://intrastat.lrmuitine.lt/

The commit also add the code region for Lithuania as well as the possibility
to set the region in the settings for that regard, i.e. the intrastat export.

task-3248669

closes odoo/enterprise#54841

X-original-commit: d7db356bffcd2619cdc4f80c6f47001a0a80f4fd
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Simon Morais Da Cunha (smdc) <smdc@odoo.com>
Co-authored-by: Habib (ayh) <h4818@users.noreply.github.com>