[IMP] project_enterprise: display project deadline in Gantt view
This commit displays the project start date and dealine in the gantt
view of its tasks.

Until this commit, the milestones linked to the tasks of the project are
displayed in the Gantt view with a diamond on the milestone
deadline. When hovered, this diamond shows a pop-up indicating the
milestone(s) and their states.

This commit imitates this behavior to display the start date and the
deadline of the project in the Gantt view with a green/red line and a
pin on the relevant date. When one or multiple milestones occured on the
same day as the start date/deadline:
- The diamond is replaced by a circle
- The pop-over indicate, on top of the milestones name, that this is the
project start date/deadline.

A small functionnal change introduced by this commit is that only
milestones related to current tasks are displayed, instead of all
milestones that relates to projects of the current tasks.

task-2647326

Part-of: odoo/enterprise#43617