[FIX] planning: compute open shifts totals correctly
Steps:
In Planning, create an open shift from 2PM to 8PM.
Note that the pill will display "2:00 PM - 8:00 PM (6h)",
which means that 6h are allocated.
For shifts with a resource that have a fixed scheduled
(shift.resource_id.calendar_id != False),
this value is computed by finding the intersection of
its calendar and the shift.
However, for shift with resource that have a flexible schedule
(shift.resource_id.calendar_id == False) or,
in this case, open shifts (shift.resource_id == False), this
value is just the duration of the shift,
maxed to the company calendar's "hours per day".

Issue:
The "Total" row's value is "03:00",
where it should have the same value as the shift.

Causes:
- `gantt_resource_work_interval` returns a dict where "no resource"
is not considered as flexible hours. As explained earlier, it should.
- When we want to access this dict on client-side, we don't handle
the case where `record.resource_id` is false, rather than a tuple.

task-3412654

closes odoo/enterprise#50473

X-original-commit: 7b8b9736ef1e42eaa558d0a83b3599e72a462850
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>