[FIX] social_youtube: display a proper error message
Before this commit
==================
At the time of the upload failure we display the standard error msg,
Because of this, user will try to upload the video again and again
and never know the exact reason for the failure.

After this commit
=================
This commit add some validation so that we avoid any upload
failure from youtube side and display a proper error msg.

1).Checks that Title is not more than 100 characters.

2).Checks that Description is not more than 5000 characters.

3).Checks that Title does not contains > or < forbidden character.

4).Checks that description does not contain > or < forbidden character.

These checks are done at time of the upload video and also at the time of the
post.

Task-3392167

closes odoo/enterprise#60579

X-original-commit: 2af7ab650e694f9b556db6d7b45e1429f6db9e50
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Naman Shah (nash) <nash@odoo.com>