[FIX] web_cohort: variable collision issue
The code for web_cohort runs several read_group calls with domains that
are dynamically generated.

It appears that a name collision found its way into this code at its
inception and has survived undetected until now, despite affecting the
returned data significantly. The problematic variable is named `domain`
and is meant to indicate the domain used for a specific cell of the
cohort, but its generic name made it leak to the base of the
computation, polluting it entirely.

This commit renames the problematic variable and adds a test to avoid
regression on that front.

closes odoo/enterprise#57116

X-original-commit: 092dc9fa8abbfd65416ca515d16ddaa8298d421d
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>