[FIX] appointment: fix linked resource availability
Don't take into account linked resources that don't have remaining
capacity or that are not part of the initial resources available.

This was leading to some errors when validating a time slot selected
because some resources with no remaining capacity where added from
resources they were linked to.

task-3424035

X-original-commit: d55041d28f1292046877df0f21fcf912c2c66255
Part-of: odoo/enterprise#49883