[IMP] project_enterprise: progress_bar improvemnts
Before this commit, the progress bar values were calculated based on the
task's start and end dates. This method proved inefficient for task
planning, as it didn't account for the possibility of a user being
assigned to multiple projects concurrently. For example, if a user had
a 40-hour work week, divided between 10 hours on Project A, 15 hours on
Project B, 5 hours on Project C, and 10 hours on Project D, the
previous system would inaccurately indicate resource overload and
potential task conflict.

With this latest update, progress bar computations are now based on the
actual allocated hours, providing a more accurate reflection of task
distribution. Additionally, if a task's end date extends beyond the
Gantt view's boundaries, the progress bar values are adjusted to
represent only the visible portion of the task, which aligns with the
percentage of the task that is actually visible.

Moreover, overlapping warning are now computed based on allocated hours
as well. The condition for a conflict is that the combined allocated
hours are larger than the workable hours of the user the planned dates.
This modification does not apply the field service tasks :

    Unlike other tasks, fsm tasks overlap in two scenarios:
        1. When the combined allocated hours of an fsm task and
            a normal task exceed the user's workable hours.
        2. When two fsm tasks have overlapping planned dates.

    Example:
        - Task A (normal) conflicts with Task B (fsm) if their combined
        hours > user's workable hours. Both have 1 conflict.
        - Introduce Task C (fsm) with no allocated hours (no conflict
        with Task A) but same time period as Task B.

        Result => Task A has 1 conflict, Task B has 2 conflicts, Task C
        has 1 conflict.

taskid:3349378

closes odoo/enterprise#44881

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>