[IMP] appointment: group by partner ids
Currently the gantt view for user appointments (dentist, interview,...)
is grouped by organizer of the event.

While this works for the purpose of managing appointments it also hides
the fact that the partners linked to each user may be attending
some other event at the same time.

To make this more clear we enable grouping by partner_ids
and we make the gantt view default to that grouping.

task-3452277

Part-of: odoo/enterprise#45114