[FIX] account_accountant: reconciliation responsiveness
This commit fixes the mobile responsiveness inside the reconciliation in
`account_accountant`.

The width of the elements was too small and the `overflow` value broke
the layout on mobile view.

The point is to make the data at least readable on mobile, but this could
be improved in the near future (Less data shown to user, displayed another
way like a list-group, etc).

task-3358275
Related to task-3326263

closes odoo/enterprise#65204

X-original-commit: 37c2170e50a7598501be44874704d7053d4106b1
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>