[FIX] mrp_workorder_hr{,_account}: set wo employee costs
Currently, the `employee_cost` of a `mrp.workcenter.productivity` is
recomputed each time we try to access its `total_cost`.

This happens because while they share the same compute method:
- `employee_cost` is stored
- `total_cost` is NOT stored

Which means that everytime `total_cost` is accessed, it will be
recomputed, re-setting the stored value of the `employee_cost`.
Overall a bad practice to mix both stored and non-stored fields in the
same compute method.

closes odoo/enterprise#64481

X-original-commit: 2527513e7090f3a0d75a7777f339e712c5ab0dda
Related: odoo/odoo#169124
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>