[IMP] l10n_pl: Update of the polish accounts, groups and positions
Currently, the polish chart of account is not up to date and is therefore
underused by user since it is not relevant.

This PR goes with a community PR which goal is to update the chart of accounts
and account groups. Because of this change, financial reports have to be
updated too so they match the new chart.

Adding this, the financial reports should be relevant and
match the new chart of accounts.

task-3332731

closes odoo/enterprise#52874

Related: odoo/odoo#146452
Signed-off-by: Quentin De Paoli <qdp@odoo.com>