[IMP] website_helpdesk: improve the routs for better SEO
Before this commit, clicking on the 'Help' menu would lead to a submit ticket
page with a  specific link.  And, accessing the '/helpdesk' route would also
display the same page. It will have a negative impact on SEO

In this commit,  if there are multiple teams, clicking on the 'Help' menu will
redirect users to a page where they can select the specific team they wish
to contact, and clicking on a team will take users to the submit ticket page.

task-3453870

closes odoo/enterprise#46282

Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>