[FIX] industry_fsm(_sale): display subtask customer when unset on parent
Steps:
Create a non-fsm task T.
Make sure it has a Customer.
Go to sub-tasks tab, and display the column "Customer".
Create a subtask and make sure it has a customer too.
Unset T's customer.

Issue:
"Customer" column disappeared on the sub-tasks tab.

Cause and Fix:
In that commit https://github.com/odoo/enterprise/pull/41700/files,
the best way to fix the bug was to hide the column.
However, we can limit this inconvenience to Field Service.

task-3462486

closes odoo/enterprise#50443

X-original-commit: 7fa97fe11389ef53a4573dc3c7d3ecd07e5172ea
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>