[IMP] l10n_fr_reports: export FEC file from General Ledger
The French FEC file was a separate module which was using the base64 enconding in order to generate the needed file.

This could take very long on DBs with a few millions AML.

Moved the generation of the report to enterprise, using an empty hook in Community, and adapted it so it uses the export_file functions which are then available and much more efficient.

Related to :
https://github.com/odoo/odoo/pull/139847
https://github.com/odoo/upgrade/pull/5435
task-3447342

closes odoo/enterprise#54113

Related: odoo/upgrade#5435
Related: odoo/odoo#149260
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>