[FIX] helpdesk: fixes generate coupon performance
Before this commit, the coupon generation in helpdesk took many seconds
because a mail for the coupon generated is sent twice (when the coupon is
created and when the user send it to the customer).

This commit improves the performance of the coupon generation in helpdesk
by avoiding sending the email when the coupon is created since the user
will send it.

task-3460049

closes odoo/enterprise#50494

X-original-commit: 5946a8d5a39fc307f964936c5e02dd5be6d55367
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>