[ADD] l10n_{ae,hk}_reports: default deferral accounts
We want to set up sensible default accounts for the deferred expenses
and deferred revenues in different localizations.

For the United Arab Emirates and Hong Kong there were no reporting
modules yet, so we are adding them here with only the setting of the
default accounts for deferred revenues and expenses.

task-3419066

closes odoo/enterprise#46383

Related: odoo/odoo#132696
Signed-off-by: William André (wan) <wan@odoo.com>