[IMP] worksheet: open the worksheet template view when closing studio
This commit's purpose is to open the worksheet template view instead of
the worksheet form view when leaving studio.

Why:
The current behavior is the default one, when leaving studio, the view
is the one that was being edited by studio.

Solution:
Create a patch for the studio navbar in order to bypass that behavior
and open the view we need in our specific case.

Step to reproduce:
- Open fsm app
- Open the configuration menu
- Click on 'worksheet template'
- Click on 'design new template'
- Close studio

The view is the form view of the worksheet, while we'd like it to be the
worksheet template form view.

Details of the implementation:
Some of the effective part of the changes were made in the worksheet module
instead of the industry_fsm_report one. The reason is that it is a
behavior that might be needed in the future for other use case (e.a
in the quality_mrp_workorder_worksheet module).
The navbar patch has been created to overwrite part of the close() method.
When our use case is triggered instead of the general one, we force the view
we need to be opened instead of the default one, then we mimic the behavior
of the close method (by removing studio from the context among other things)
The context of the action of the template is changed for the tasks at
its creation and not only when the get_x_form_action method is called.
The reason for this change is that when the studio view is reloaded, the
context is discarded, and the default behavior is applied again. By adding the
2 extra keys in the context of the action at its creation (and thus in the db),
we ensure that these context keys will be present on the reload.
The tour test was added in industry_fsm_report since the fsm module is
needed to execute the use case.

task-3316535

closes odoo/enterprise#46874

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>