[FIX] web_studio: make SelectMenu value optional for options
This commit allows the value of an option to be unset. Before this commit,
once a value had been set, it was impossible to remove it from the view.
For example, an image size couldn't be unset to keep the default behavior.

A test has been added to verify that the SelectMenu has a remove button, and
that the value can be remove effectively.

task-3470048

closes odoo/enterprise#52126

X-original-commit: 0ec5fe8d9d24563c0458f15b4354b88f9b10546a
Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>