[FIX] web_studio: make SelectMenu value optional for options
This commit allows the value of an option to be unset. Before this commit,
once a value had been set, it was impossible to remove it from the view.
For example, an image size couldn't be unset to keep the default behavior.

A test has been added to verify that the SelectMenu has a remove button, and
that the value can be remove effectively.

task-3470048

closes odoo/enterprise#52034

X-original-commit: 8e49932b26a26cfba3774584874c4d6cd7b6d3db
Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>