[FIX] knowledge: fix burger menu overflow when resizing sidebar
Currently, the burger menu of the Knowledge editor can overflow the
layout when the user resizes the sidebar on the desktop view and
switches to the mobile view.

Steps to reproduce the issue:
1. Open the Knowledge editor on a desktop.
2. Resize the sidebar with the resizer.
3. Reduce the size of your window to switch to the mobile view.
4. Open the burger menu.

=> The buger menu keeps the width set after the resize. In some cases,
the burger menu can be too small or can overflow the container.

TO BE: The burger menu should not be affected by the resize. It should
have a width of 100% minus an offset of 50px.

This commit will address that issue by defining a css variable for the
sidebar width. The Javascript will update that variable when the user
use the resize handler. The CSS will then use a media query to constraint
the sidebar to have that width on desktop devices.

task-3507172

X-original-commit: 18cd035924ce5fb9b3c0ecccb77b8ab7adca7c1e
Part-of: odoo/enterprise#51280