[FIX] web_studio: renaming a new field alters its description in DB
In studio, add a new field in a form view.
Assign a new label to that field. Since the field is new and guaranteed to not contain data,
its technical name can be changed alongside its label.

Before this commit, the field's technical name was changed, then an operation was passed to modify the string attribute in the view and modify the field's description on ir.model.fields

This worked but was impractical in terms of translations. When translating a field's description, the
string in the view was not translated, and could not be so.

In this flow, we assume that the user want to modify the properties of the field itself (ir.model.fields)
since even the technical name is changed.

After this commit, when renaming a field that way, we pass the new label and write it along. At that point we do nothing in the view itself, to let the own property ir.model.fields:description do its generic work of labelling a field

opw-3500806
opw-3610918

closes odoo/enterprise#53700

X-original-commit: 296d17929ef1387f1ea1ddf378d2800747cb9829
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>