[FIX] whatsapp: include archived templates when synchronizing
Steps to reproduce:
1. Install the Whatsapp module.
2. archive any of the submitted templates.
3. now go to WhatsApp > configuration > WhatsApp business account then sync the
   template
It creates a new template instead of updating the archived template.

Reason:
In the sync a search on templats is performed and archived templates are
automatically excluded.

Solution:
Search archive template also in sync template. We also fix template status
update webhooks for the same kind of fix.

Task-3516448
Sub-part of Task-3608129 (Whatsapp: Moultifix !)

X-original-commit: c9194e3ab363aba6e2821ad6a9a0303de43598ad
Part-of: odoo/enterprise#52840