[FIX] web_studio: allow removal of conditional invisible attribute
Consider a field in a form view having a conditional invisible property. When
toggling the invisible property (setting it to either true or false), the
onChangeModifier method removes the invisible property from the attrs object.

```xml
<!-- in arch -->
<field name="name" attrs="{'invisble': '[]'}" />

<!-- expected after setting invisible to false -->
<field name="name" attrs="" />
```

Before this commit, the attrs object was sent empty, and the server intepreted it as not changing
relative to the original arch. In practice, the user's request failed, because the field was still marked as invisible.

After this commit, we send empty keys in attrs that should be removed, hence, the server interprets it as changing and applies the new values.

opw-3497068

closes odoo/enterprise#52589

X-original-commit: b1628b4e5ed850d4d8b476636eeb3317c5b1d483
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>