[FIX] web_studio: allow removal of conditional invisible attribute
Consider a field in a form view having a conditional invisible property. When
toggling the invisible property (setting it to either true or false), the
onChangeModifier method removes the invisible property from the attrs object.

```xml
<!-- in arch -->
<field name="name" attrs="{'invisble': '[]'}" />

<!-- expected after setting invisible to false -->
<field name="name" attrs="" />
```

Before this commit, the attrs object was sent empty, and the server intepreted it as not changing
relative to the original arch. In practice, the user's request failed, because the field was still marked as invisible.

After this commit, we send empty keys in attrs that should be removed, hence, the server interprets it as changing and applies the new values.

opw-3497068

closes odoo/enterprise#52675

X-original-commit: 7b6ae0b37153fcec0cf3ce95b29c1784c1f4ff6d
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>