[MOV] l10n_de_reports: Move code for setting default tags on accounts
Previously, the code that sets default tags on bank accounts was in
l10n_de_reports, so was not yet loaded when the demo company is created
at module init of l10n_de.

As a result, the demo company was created without the necessary tags
on the 1001 Cash and 1201 Bank accounts, which meant that the Balance
Sheet would not be impacted by these accounts on the demo company.

This commit fixes this.

taskid:none

closes odoo/enterprise#52302

X-original-commit: 7e67d736b53c261ed751a5bcdc1b44651e9234b2
Related: odoo/odoo#145357
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>