[FIX] appointment: crash on cancel if extra attendees
Steps to reproduce
===================
Step 1. As a customer, schedule a meeting
Step 2. Add extra attendees from the backend through the calendar event
Step 3. Now again go to the appointment confirmation page and try to cancel the
meeting.
-> error(page crash)

Technical
==========
If there are more than 2 attendees in the meeting then the person who is
cancelling the meeting should only leave the meeting and his leaving
notification should be display on the chatter by the organizer.

we are removing the attendee before and then trying to send the message from
that attendee, so it is giving a traceback of not finding the record.

After this commit
=================
Now the traceback is gone and user can cancel the meeting.

Task-3505659

closes odoo/enterprise#51327

X-original-commit: 9026cea86d43cdac16bb8f60ca176e62737b28ee
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>