[FIX] knowledge: ensure article content is saved on visibility change
Before this commit:
Changing the visibility of a Workspace article did not save the content,
resulting in content loss.

Reason:
The issue was caused by the behavior of the `resetEditor` function. From
`_onChangeVisibility`, `resetEditor` function triggers after calling `load`
function, which resulted in the unintentional removal of the article content.

After this commit:
Now article content is saved when visibility changes.

Task-3491577

closes odoo/enterprise#50639

X-original-commit: c11a183c8d7c45279499078d2db99a359eba6156
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>