[IMP] documents_project: change portal default task groupby
Since 16.0, when viewing the tasks of a project with shared documents in the portal,
the tasks are grouped by project to allow displaying the "Documents" button.

However, it doesn't make much sense to group by project if there is only a single project.

This commit changes the default groupby to be by stage, like for all the other projects,
and adds the button next to the first stage instead.

Community: https://github.com/odoo/odoo/pull/136126

Task-3438493

closes odoo/enterprise#47703

Related: odoo/odoo#136126
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>