[FIX] l10n_be_hr_payroll_fleet: fix display of fuel type for new car
https://github.com/odoo/enterprise/pull/34883 started showing the fuel type of
the car_id in the form view of hr_contract. However, when selecting new cars,
the car_id won't be set. Instead we'll set the new_car_model_id field and the
fuel type won't be shown. This commit changes the fuel_type field in
hr_contract to be a computed field that will either take the fuel_type from
car_id or from new_car_model_id.

task-3515936

closes odoo/enterprise#50698

X-original-commit: a4ba581d0490639aa1c6cc38a0aad3c0e1b2743b
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Leonardo Pavan Rocha <lpr@odoo.com>