[FIX] worksheet: update name in qweb arch
Steps:
In the default worksheet template,
add the field "ID" and change its label as "Proute".
Go to a helpdesk ticket, and create a worksheet.
See that the label is "Proute".
Now share it with a portal user.
Log in as that user, and go on the ticket.

Issue:
The label of the worksheet's field is "ID".

Cause:
The back-end template uses `<field name="id" string="Proute">`.
It is then translated to a qweb as `<div string="Proute">ID</div>`
because it sets the name of the field as the description.

Fix:
Set the string of the tag if any, or else the name of the field.

task-3479545

closes odoo/enterprise#59823

X-original-commit: 67e7c6ba47229f2d952e33e8c06c5c78e697045d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>