[IMP] industry_fsm: fix default user get assignee when creating new task
Before this commit, When the user creates a new task, the assignee
gets automatically set to the current user. This behavior, however,
was inconsistent with how other menus in the Field Service module
operated.

After this commit, The assignee won't be set when creating a task. it
will remain empty.

task-3475473

closes odoo/enterprise#47752

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>