[FIX] web_gantt: avoid traceback when resizing browser
Steps
=====
- Open app Project
- In settings of Project activate task dependencies
- Open a project with dependencies between task (e.g. research and
development in demo data)
- Go to the Gantt view of the tasks of this project in mobile view (or
with as small window)
- Deactivate all filters to ensure that tasks with dependencies appear
in the current view
- Switch to desktop view (or resize the window)

Issue
=====
A traceback appears indicating "Cannot read properties of
undefined (reading '__pill__1')".

Cause
=====
The connections between dependent pill is only displayed if the user is
not in mobile view. To avoid unecessary memory allocation, some variable
used to manage the connections display (e.g. mappingPillToConnectors),
are only declared if the user is not in mobile view. However, when
resizing the window (or leaving mobile view), the onWillRender method is
called and can make use of those variables if the connections need to be
displayed. This create the traceback are those variables have not been
declared.

Fix
===
The method computeDerivedParams used to manage such variable declaration
is called in the onWillRender method under the condition that connectors
were not present in the previous render (i.e. it was in mobile view).

closes odoo/enterprise#51574

X-original-commit: 9d5857bddced3881607a0c932a5276a694b8037b
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>