[IMP] sale_timesheet_enterprise: take all timesheets into account in portal
Before this commit, when the portal user has no access to one subtask then the
totals do not correspond to the reality (as we have in the backend views).

This commit computes the real totals even if the portal user has no access to
some sub tasks to be sure the totals displayed in the parent task correspond to
the reality.

Related PR: odoo/odoo#128407

task-3423713

closes odoo/enterprise#48150

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>