[IMP] web_grid: focus on anchor
In grid view, when the range is "month", the last column-days are not
visible: you have to scroll left to see them.
With this commit, when today (or its equivalent date on whatever month
you're looking at) is at the end of the month, we automatically scroll
to make it visible.

task-3483742

closes odoo/enterprise#49603

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>