[ADD] pos_pricer: pos_pricer module added
This PR adds a new model "pos_pricer"
It allows integration of Odoo with Pricer.

Users can now associate their products in Odoo to
Pricer electronic price tags, which allow them
to display product information and price

Every 12 hours in a cron we check if the products with Pricer
tags linked to them were modified and if so, we
call Pricer API to modify the displayed product information. (this can
also be done manually)

task-3414052

closes odoo/enterprise#52186

X-original-commit: 99fd536e27bbc08d3caa919d0d0a51fa1727645c
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>