[IMP] helpdesk_sale_timesheet,industry_fsm_sale,timesheet_grid: add various improvements
Before this commit:
- confirm time spent wizard: the implication of each button
from the is not explicit.
- it makes no sense to timesheet on a down payment as this
is not the service that was sold to the customer.
- the employee/SOL mapping tip is not accurate in field service
as the behavior is a bit different there.

After this commit:
- confirm time spent wizard: buttons changed:
save -> log time
discard -> resume timer
delete -> fa-trash icon
'discard' and 'delete' buttons are switched
- it's not possible to select an SOL linked to a down payment
for projects, tasks, timesheets,
- project.project form view > invoicing notebook > the tip is
different for fsm projects.

task-3476795

closes odoo/enterprise#48396

Related: odoo/odoo#137570
Related: odoo/upgrade#5377
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>