[FIX] l10n_ec_edi: fix base amount computation of withhold line
Steps to reproduce:
- Install l10n_ec_edi
- Create a vendor bill:
  * Customer: [Ecuadorian contact] (e.g. Instituto Ecuatoriano de Seguridad Social)
  * Payment Method (SRI): Otros con utilización del sistema financiero
  * Document Number: [any]
  * Invoice Lines:
    1. Product: [any product] - Price: 100 - Taxes: IVA 12%
    2. Product: [no product] - Price: 0 [*] - Taxes: IVA 12%
    [*] A price of 0 is used to highlight the issue, but the amount doesn't matter
- Save and confirm the invoice
- Click on "ADD WITHHOLD" button
=> The base amount of the withhold lines is not computed correctly.
One line with a retention tax seems correct: (it's only correct because the tax amount
of the second invoice line is 0.00)
* Tax Support: 01 Tax credit for VAT declaration (services and goods other than inventories and fixed assets)
* Tax: 10% Retención IVA
* Base: 12.00
* Amount: 1.20
The second line with a retention tax is not correct:
* Tax Support: 01 Tax credit for VAT declaration (services and goods other than inventories and fixed assets)
* Tax: 20% Retención IVA
* Base: 12.00
* Amount: 2.40
Base and Amount of the second line should be 0.
This line doesn't appear if the invoice line without product and a price of 0 is removed.

Cause:
During the computation, the subtotals and tax amounts of each invoice lines are summed
and grouped by Tax Support code.
The Base field of the withhold line is then computed from these grouped subtotals or
taxes, depending only on the Tax Support code, without taking into account the fact that
the withhold tax can be different.
That's why the base amount of the incorrect withhold line is 12 (instead of 0).
As it has the same Tax Support code than the first line, its value is fetched from
the sum made for that Tax Support code.

Solution:
The computation of the Base amount of the withhold lines should take into account
the Tax Support code, but also the applied tax.

opw-3478803

closes odoo/enterprise#51710

X-original-commit: d63c1367ad97ee4ab6f08f3f04bd35895a1f6955
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>