[IMP] hr_payroll: improve basic amount calculation in payslips.
This commit improves the functionality of payslips by incorporating the hourly
wage into the basic amount on the payslip. This enhancement ensures that if
an hourly wage is available in the contract, it is appropriately factored into
the basic amount calculation on the payslip.

Backward-Port-Of: https://github.com/odoo/enterprise/pull/48510
X-original-commit: https://github.com/odoo/enterprise/pull/48510/commits/9b066542e4aa59dec724a83f56e5cb4f9dd8eb83

Task-3514674

closes odoo/enterprise#53375

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>