[IMP] sale_subscription: allows to cancel order, reopen the parent and fix logs
Before this commit, when an order was canceled, if it was a renewal, the parent
order remained closed. This commit allows to reopen the parent order. It will be
useful if the order was canceled by mistake. It also clean the logs to avoid
displaying churned contrat that should not appear in the statistics.

Taskid: 3527957
Part-of: odoo/enterprise#51578