[FIX] documents_project_sale: unlinked to-do stage
Steps to reproduce:
-> Install documents_project_sale module.
-> Open project
-> Go to  configuration > task stages.
-> There is a 'To Do' stage that isn't linked to any project.

Issue:
before this commit, the 'To Do' stage that stage that isn't linked to any project.
which is showing abnormally.

Cause:
The created stage will be removed automatically but the problem is it searches for
'To do' not 'To Do' that's why it was not removed.

Fix:
after this commit, when we go to the task stages it will search for the 'To Do'
stage if there is no project linked to the 'To Do' stage then it will be removed
automatically the 'To Do' stage.

task-3550638

closes odoo/enterprise#54635

X-original-commit: 003d6c389072345685b568fce33ddaca79dab807
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>